fixed not being able to add products to quotes

+++ src/main/java/com/bpatel2001/invoicing_system/dao/QuoteRepository.java
@@ -1,9 +1,10 @@
 package com.bpatel2001.invoicing_system.dao;
 
-import com.bpatel2001.invoicing_system.entity.Quotes;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
+import com.bpatel2001.invoicing_system.entity.Quotes;
+
 @RepositoryRestResource(path="quotes")
-public interface QuoteRepository extends JpaRepository<Quotes, Long> {
+public interface QuoteRepository extends JpaRepository<Quotes, Integer> {
 }
